data_pipeline/requirements.txt
pandas==2.0.3
apache-beam[gcp]
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
requests==2.31.0
dill==0.3.1.1
pytz